+++ jobslib/__init__.py
@@ -46,15 +46,13 @@ inherit and override :class:`Config` and/or :class:`Context` and set
 from .cmdlineparser import argument
 from .config import Config, ConfigGroup, option
 from .context import Context, cached_property
-from .metrics import BaseMetrics
 from .oneinstance import OneInstanceWatchdogError
 from .tasks import BaseTask
 from .version import VERSION
 
 __all__ = [
     'argument', 'Config', 'ConfigGroup', 'option', 'Context',
-    'cached_property', 'BaseMetrics', 'OneInstanceWatchdogError',
-    'BaseTask'
+    'cached_property', 'OneInstanceWatchdogError', 'BaseTask',
 ]
 
 __version__ = VERSION