+++ jobslib/metrics/influxdb.py
@@ -1,14 +1,17 @@
+"""
+Module :mod:`jobslib.metrics.influxdb` provides :class:`InfluxDBMetrics`
+writer.
+"""
 
-import os
-
-from objectvalidator import option
+import datetime
+import logging
+import time
 
 from . import BaseMetrics
-from ..config import ConfigGroup
-
-from influxdb_wrapper import DoporucovaniInfluxDBCLient
 
 __all__ = ['InfluxDBMetrics']
+
+logger = logging.getLogger(__name__)
 
 
 class InfluxDBMetrics(BaseMetrics):
@@ -19,79 +22,42 @@ class InfluxDBMetrics(BaseMetrics):
 
     .. code-block:: python
 
+        INFLUXDB = {
+            'host': 'hostname',
+            'port': 8086,
+            'username': 'root',
+            'password': 'root',
+            'database': 'dbname',
+        }
+
         METRICS = {
             'backend': 'jobslib.metrics.influxdb.InfluxDBMetrics',
-            'options': {
-                'host': '127.0.0.1',
-                'port': 8086,
-                'username': 'root',
-                'password': 'root',
-                'database': 'dbname'
-            }
         }
-
     """
 
-    class OptionsConfig(ConfigGroup):
-        @option(required=True, attrtype=str)
-        def host(self):
-            """
-            InfluxDB host
-            """
-            host = os.environ.get('JOBSLIB_METRICS_INFLUXDB_HOST')
-            if host:
-                return host
-            return self._settings.get('host', '127.0.0.1')
-
-        @option(attrtype=int)
-        def port(self):
-            """
-            InfluxDB port
-            """
-            port = os.environ.get('JOBSLIB_METRICS_INFLUXDB_PORT')
-            if port:
-                return int(port)
-            return self._settings.get('port', 8086)
-
-        @option(required=True, attrtype=str)
-        def username(self):
-            """
-            InfluxDB username
-            """
-            username = os.environ.get('JOBSLIB_METRICS_INFLUXDB_USERNAME')
-            if username:
-                return username
-            return self._settings.get('username', "root")
-
-        @option(required=True, attrtype=str)
-        def password(self):
-            """
-            InfluxDB password
-            """
-            password = os.environ.get('JOBSLIB_METRICS_INFLUXDB_PASSWORD')
-            if password:
-                return password
-            return self._settings.get('password', "root")
-
-        @option(required=True, attrtype=str)
-        def database(self):
-            """
-            InfluxDB database
-            """
-            database = os.environ.get('JOBSLIB_METRICS_INFLUXDB_DBNAME')
-            if database:
-                return database
-            return self._settings.get('database', "dbname")
-
-    def __init__(self, context, options):
-        super().__init__(context, options)
-        kwargs = {
-            k: v for k, v in options.as_kwargs.items()
-            if v is not None
-        }
-
-        self._client = DoporucovaniInfluxDBCLient(
-            app=context.config.task_class.name, **kwargs)
-
-    def push_monitoring_metrics(self, metrics):
-        pass
+    def push(self, metrics):
+        current_dt = datetime.datetime.utcfromtimestamp(time.time())
+        ts = current_dt.strftime('%Y-%m-%dT%H:%M:%SZ')
+        task_name = self.context.config.task_class.name
+        try:
+            points = []
+            for metric_name, metric_value in metrics.items():
+                tags = {
+                    'task': task_name,
+                }
+                for k, v in metric_value.get('tags', {}).items():
+                    if k in tags:
+                        raise Exception("Tag '{}' is reserved".format(k))
+                    tags[k] = v
+                metric = {
+                    'measurement': metric_name,
+                    'tags': tags,
+                    'time': ts,
+                    'fields': {
+                        'value': metric_value['value'],
+                    },
+                }
+                points.append(metric)
+            self.context.influxdb.write_points(points)
+        except Exception:
+            logger.exception('Push monitoring metrics into InfluxDb failed')